move actix-specific stuff from lib into server so only server has to compile it
