+++ oxen-rust/src/lib/src/api/client/file.rs
@@ -207,7 +207,7 @@ pub async fn delete_file(
 #[cfg(test)]
 mod tests {
 
-    use actix_web::web::Bytes;
+    use bytes::Bytes;
 
     use crate::constants::DEFAULT_BRANCH_NAME;
     use crate::error::OxenError;
