+++ oxen-rust/src/lib/src/constants.rs
@@ -270,6 +270,3 @@ pub fn chunk_size() -> u64 {
         AVG_CHUNK_SIZE
     }
 }
-
-// Oxen request Id
-pub const OXEN_REQUEST_ID: &str = "x-oxen-request-id";
